Wrapping Up Version 1.0
Modified VPN state from Boolean to a tri-state, to handle a partially-enabled state.

Changed the icon from a lock to a pan.

Disabled advertised shortcuts.

Specified in the setup that the components are not 64-bit.

Removed separate x86 and x64 configurations from setup.

Removed the checkboxes from the UI, in favor of the buttons. Enabled the functionality for the buttons.